+++ src/Commands/Core/FixStructureCommand.php
@@ -96,7 +96,7 @@ class FixStructureCommand extends Command
 
             $io->block('saving translations of this set...');
 
-            $storageSaver = StorageFactory::getStorage($set);
+            $storageSaver = StorageFactory::getInstance()->getStorage($set);
 
             $storageSaver->saveTranslationSet($set);
         }