+++ src/Commands/Exchange/ImportCommand.php
@@ -100,7 +100,7 @@ class ImportCommand extends Command
             }
 
             # get correct storage saver from our current set
-            $storageSaver = StorageFactory::getStorage($set);
+            $storageSaver = StorageFactory::getInstance()->getStorage($set);
 
             # build the correct importer for our exchange format
             # and pass on the matching storage saver of our current ste